refactor(add_new_goal): consolidate providers and move create goal logic
- Replace GoalResourcesProvider with AddNewGoalProvider that handles both resources and goal creation
- Move create goal button logic from form to app bar
- Update all references to use the new provider
- Remove unused CreateNewGoalButton widget
