+++ lib/features/add_new_goal/view/screens/add_new_goal_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:goal_tree/features/add_new_goal/providers/goal_resources_provider.dart';
+import 'package:goal_tree/features/add_new_goal/providers/add_new_goal_provider.dart';
 import 'package:goal_tree/features/add_new_goal/view/widgets/add_new_goal_app_bar.dart';
 import 'package:goal_tree/features/add_new_goal/view/widgets/add_new_goal_form.dart';
 import 'package:provider/provider.dart';
@@ -9,14 +9,14 @@ class AddNewGoalScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: const PreferredSize(
-        preferredSize: Size.fromHeight(kToolbarHeight),
-        child: AddNewGoalAppBar(),
-      ),
-      body: ChangeNotifierProvider(
-        create: (context) => GoalResourcesProvider(),
-        child: const AddNewGoalForm(),
+    return ChangeNotifierProvider(
+      create: (_) => AddNewGoalProvider(),
+      child: const Scaffold(
+        appBar: PreferredSize(
+          preferredSize: Size.fromHeight(kToolbarHeight),
+          child: AddNewGoalAppBar(),
+        ),
+        body: AddNewGoalForm(),
       ),
     );
   }
