+++ lib/features/add_new_goal/view/widgets/add_new_goal_form.dart
@@ -1,67 +1,39 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:goal_tree/core/models/goal_model.dart';
-
-import 'package:goal_tree/core/models/goal_priority_enum.dart';
 import 'package:goal_tree/core/theme/app_text_styles.dart';
 import 'package:goal_tree/features/add_new_goal/model/validators.dart';
-import 'package:goal_tree/features/home/cubit/home_cubit.dart';
 import 'package:provider/provider.dart';
-import 'package:goal_tree/features/add_new_goal/providers/goal_resources_provider.dart';
+import 'package:goal_tree/features/add_new_goal/providers/add_new_goal_provider.dart';
 import 'package:goal_tree/features/add_new_goal/view/widgets/add_new_goal_custom_text_form_field.dart';
 import 'package:goal_tree/features/add_new_goal/view/widgets/add_resource_button.dart';
-import 'package:goal_tree/features/add_new_goal/view/widgets/create_new_goal_button.dart';
 import 'package:goal_tree/features/add_new_goal/view/widgets/priority_list.dart';
 import 'package:goal_tree/features/add_new_goal/view/widgets/resources_list.dart';
 import 'package:goal_tree/features/home/view/widgets/add_resource_form.dart';
 import 'package:goal_tree/features/home/view/widgets/cancel_adding_resource_button.dart';
 
-class AddNewGoalForm extends StatefulWidget {
+class AddNewGoalForm extends StatelessWidget {
   const AddNewGoalForm({super.key});
 
   @override
-  State<AddNewGoalForm> createState() => _AddNewGoalFormState();
-}
-
-class _AddNewGoalFormState extends State<AddNewGoalForm> {
-  final _formKey = GlobalKey<FormState>();
-  final TextEditingController goalTitle = TextEditingController();
-  final TextEditingController goalDescription = TextEditingController();
-  final TextEditingController goalDeadlineController = TextEditingController();
-  DateTime? goalDeadline;
-  GoalPriorityEnum goalPriority = GoalPriorityEnum.medium;
-  final TextEditingController goalNotes = TextEditingController();
-  bool addingResource = false;
-  bool _isCreatingGoal = false;
-
-  @override
-  void dispose() {
-    goalDeadlineController.dispose();
-    goalTitle.dispose();
-    goalDescription.dispose();
-    goalNotes.dispose();
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext context) {
-    return Consumer<GoalResourcesProvider>(
+    final provider = Provider.of<AddNewGoalProvider>(context);
+    return Consumer<AddNewGoalProvider>(
       builder: (context, goalResourcesProvider, child) {
         return Form(
-          key: _formKey,
+          key: provider.formKey,
           child: SingleChildScrollView(
             padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
                 AddNewGoalCustomTextFormField(
-                  controller: goalTitle,
+                  controller: provider.goalTitle,
                   hintText: 'Goal Title',
                   validator: Validators.titleValidation,
                 ),
                 const SizedBox(height: 24),
                 AddNewGoalCustomTextFormField(
-                  controller: goalDescription,
+                  controller: provider.goalDescription,
                   hintText: 'Goal Description',
                   maxLines: 5,
                 ),
@@ -69,36 +41,21 @@ class _AddNewGoalFormState extends State<AddNewGoalForm> {
                 Text('Deadline', style: AppTextStyles.headText).tr(),
                 const SizedBox(height: 24),
                 AddNewGoalCustomTextFormField(
-                  controller: goalDeadlineController,
+                  controller: provider.goalDeadlineController,
                   hintText: 'Select Deadline',
                   suffixIcon: const Icon(Icons.calendar_month),
                   readOnly: true,
-                  onTap: () async {
-                    DateTime? pickedDate = await showDatePicker(
-                      context: context,
-                      initialDate: DateTime.now(),
-                      firstDate: DateTime.now(),
-                      lastDate: DateTime(2100),
-                    );
-                    if (pickedDate != null) {
-                      goalDeadline = pickedDate;
-                      goalDeadlineController.text = DateFormat(
-                        'yyyy-MM-dd',
-                      ).format(pickedDate);
-                    }
-                  },
+                  onTap: () => provider.pickTime(context),
                 ),
                 const SizedBox(height: 24),
                 Text('Priority', style: AppTextStyles.headText).tr(),
                 const SizedBox(height: 24),
                 PriorityList(
-                  onPriorityChanged: (value) {
-                    goalPriority = value;
-                  },
+                  onPriorityChanged: (value) => provider.changePriority(value),
                 ),
                 const SizedBox(height: 24),
                 AddNewGoalCustomTextFormField(
-                  controller: goalNotes,
+                  controller: provider.goalNotes,
                   hintText: 'Notes',
                   maxLines: 5,
                 ),
@@ -106,72 +63,18 @@ class _AddNewGoalFormState extends State<AddNewGoalForm> {
                 Text('Resources', style: AppTextStyles.headText).tr(),
                 const SizedBox(height: 24),
                 const ResourcesList(),
-                if (addingResource) const AddResourceForm(),
+                if (provider.addingResource) const AddResourceForm(),
 
                 const SizedBox(height: 24),
-                !addingResource
+                !provider.addingResource
                     ? AddResourceButton(
-                        onPressed: () {
-                          setState(() {
-                            addingResource = !addingResource;
-                          });
-                        },
+                        onPressed: () => provider.toggleAddingResource(),
                       )
                     : CancelAddingResourceButton(
-                        onPressed: () {
-                          setState(() {
-                            addingResource = !addingResource;
-                          });
-                        },
+                        onPressed: () => provider.toggleAddingResource(),
                       ),
 
                 const SizedBox(height: 24),
-                CreateNewGoalButton(
-                  onPressed: _isCreatingGoal
-                      ? null
-                      : () async {
-                          if (!(_formKey.currentState?.validate() ?? false)) {
-                            return;
-                          }
-
-                          setState(() {
-                            _isCreatingGoal = true;
-                          });
-
-                          try {
-                            await context.read<HomeCubit>().addGoal(
-                              GoalModel(
-                                title: goalTitle.text,
-                                description: goalDescription.text,
-                                deadline: goalDeadline,
-                                priority: goalPriority.index,
-                                notes: goalNotes.text,
-                                initialResources:
-                                    goalResourcesProvider.goalResources,
-                              ),
-                            );
-                            if (context.mounted) {
-                              Navigator.pop(context);
-                            }
-                          } catch (e) {
-                            if (context.mounted) {
-                              ScaffoldMessenger.of(context).showSnackBar(
-                                SnackBar(
-                                  content: Text(
-                                    'Failed to add goal: ${e.toString()}',
-                                  ),
-                                ),
-                              );
-                            }
-                          } finally {
-                            if (mounted) {
-                              setState(() {
-                                _isCreatingGoal = false;
-                              });
-                            }
-                          }
-                        },
-                ),
               ],
             ),
           ),
