+++ lib/features/home/view/widgets/add_resource_form.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:goal_tree/core/models/resource_model.dart';
 import 'package:goal_tree/features/add_new_goal/model/validators.dart';
-import 'package:goal_tree/features/add_new_goal/providers/goal_resources_provider.dart';
+import 'package:goal_tree/features/add_new_goal/providers/add_new_goal_provider.dart';
 import 'package:goal_tree/features/add_new_goal/view/widgets/add_new_goal_custom_text_form_field.dart';
 import 'package:goal_tree/features/add_new_goal/view/widgets/add_resource_button.dart';
 import 'package:provider/provider.dart';
@@ -38,7 +38,7 @@ class _AddResourceFormState extends State<AddResourceForm> {
           AddResourceButton(
             onPressed: () {
               if (_formKey.currentState?.validate() ?? false) {
-                context.read<GoalResourcesProvider>().addResource(
+                context.read<AddNewGoalProvider>().addResource(
                   ResourceModel(
                     name: _resourceName ?? '',
                     link: _resourceLink ?? '',
